+++ src/main/java/com/nyakako/simplesave/controller/TransactionController.java
@@ -2,14 +2,19 @@ package com.nyakako.simplesave.controller;
 
 import com.nyakako.simplesave.model.Category;
 import com.nyakako.simplesave.model.Transaction;
+import com.nyakako.simplesave.model.User;
+import com.nyakako.simplesave.security.CustomUserDetails;
 import com.nyakako.simplesave.service.CategoryService;
 import com.nyakako.simplesave.service.TransactionService;
+import com.nyakako.simplesave.service.UserService;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDate;
 
 import org.springframework.lang.NonNull;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,25 +28,37 @@ public class TransactionController {
 
     private final TransactionService transactionService;
     private final CategoryService categoryService;
+    private final UserService userService;
 
-    public TransactionController(TransactionService transactionService, CategoryService categoryService) {
+    public TransactionController(TransactionService transactionService, CategoryService categoryService,
+            UserService userService) {
         this.transactionService = transactionService;
         this.categoryService = categoryService;
+        this.userService = userService;
     }
 
     @GetMapping("/transactions")
-    public String showTransactions(Model model) {
-        model.addAttribute("transactions", transactionService.findAllTransactions());
+    public String showTransactions(Model model, Authentication authentication) {
+        // ログインユーザーの明細取得
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
+        Iterable<Transaction> transactions = transactionService.findTransactionsByUserId(userId);
+
+        // 全明細取得
+        // List<Transaction> transactions = transactionService.findAllTransactions();
+        model.addAttribute("transactions", transactions);
         model.addAttribute("title", "取引一覧 - simplesave");
         model.addAttribute("content", "transactions");
         return "layout";
     }
 
     @GetMapping("/transactions/new")
-    public String newTransaction(Model model) {
+    public String newTransaction(Model model, Authentication authentication) {
         LocalDate today = LocalDate.now();
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
         model.addAttribute("today", today);
-        model.addAttribute("categories", categoryService.findAllCategories());
+        model.addAttribute("categories", categoryService.findCategoriesByUserId(userId));
         model.addAttribute("title", "新規明細登録 - simplesave");
         model.addAttribute("content", "new-transaction");
         return "layout";
@@ -49,31 +66,62 @@ public class TransactionController {
 
     @PostMapping("/transactions/new")
     public String addTransaction(@NonNull @ModelAttribute Transaction transaction,
-            @NonNull @RequestParam("categoryId") Long categoryId) {
+            @NonNull @RequestParam("categoryId") Long categoryId, Authentication authentication) {
+
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
+        if (userId != null) {
+            User user = userService.findUserById(userId).orElse(null);
+            transaction.setUser(user);
+        }
+
         // categoryIdを使用してCategoryオブジェクトを取得
         Category category = categoryService.findCategoryById(categoryId).orElse(null);
 
         // TransactionオブジェクトにCategoryをセット
         transaction.setCategory(category);
+
         transactionService.saveTransaction(transaction);
         return "redirect:/transactions";
     }
 
     @GetMapping("/transactions/edit/{id}")
-    public String editTransaction(@PathVariable @NonNull Long id, Model model) {
+    public String editTransaction(@PathVariable @NonNull Long id, Model model, Authentication authentication) {
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
+
         Transaction transaction = transactionService.findTransactionById(id).orElse(null);
+        // 対象が存在しない、または対象の所有者がログインユーザーでない場合はアクセス制限
+        if (transaction == null || !transaction.getUser().getId().equals(userId)) {
+            // アクセス拒否の処理
+            throw new AccessDeniedException("このページにアクセスする権限がありません。");
+        }
+
         BigDecimal amount = transaction.getAmount();
         transaction.setAmount(amount.setScale(0, RoundingMode.DOWN));
         model.addAttribute("transaction", transaction);
-        model.addAttribute("categories", categoryService.findAllCategories());
+        model.addAttribute("categories", categoryService.findCategoriesByUserId(userId));
         model.addAttribute("title", "明細編集 - simplesave");
         model.addAttribute("content", "edit-transaction");
         return "layout";
     }
 
     @PostMapping("/transactions/edit/{id}")
-    public String updateTransaction(@PathVariable Long id, @NonNull @ModelAttribute Transaction transaction,
-            @NonNull @RequestParam("categoryId") Long categoryId) {
+    public String updateTransaction(@PathVariable @NonNull Long id, @NonNull @ModelAttribute Transaction transaction,
+            @NonNull @RequestParam("categoryId") Long categoryId, Authentication authentication) {
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
+        if (userId != null) {
+            User user = userService.findUserById(userId).orElse(null);
+            transaction.setUser(user);
+        }
+
+        Transaction transactionForConfirm = transactionService.findTransactionById(id).orElse(null);
+        // 対象の所有者がログインユーザーでない場合はアクセス制限
+        if (!transactionForConfirm.getUser().getId().equals(userId)) {
+            // アクセス拒否の処理
+            throw new AccessDeniedException("このページにアクセスする権限がありません。");
+        }
         Category category = categoryService.findCategoryById(categoryId).orElse(null);
         transaction.setCategory(category);
 
@@ -82,7 +130,16 @@ public class TransactionController {
     }
 
     @PostMapping("/transactions/delete/{id}")
-    public String deleteTransaction(@NonNull @PathVariable Long id) {
+    public String deleteTransaction(@NonNull @PathVariable Long id, Authentication authentication) {
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long userId = userDetails.getUserId(); // ユーザーIDの取得
+
+        Transaction transaction = transactionService.findTransactionById(id).orElse(null);
+        // 対象が存在しない、または対象の所有者がログインユーザーでない場合はアクセス制限
+        if (transaction == null || !transaction.getUser().getId().equals(userId)) {
+            // アクセス拒否の処理
+            throw new AccessDeniedException("このページにアクセスする権限がありません。");
+        }
         transactionService.deleteTransacition(id);
         return "redirect:/transactions";
     }
